+++ src/lib/core/EditorObject.ts
