+++ src/lib/core/index.ts
@@ -1,2 +1,4 @@
 export * from './Plugin';
 export * from './Property';
+export * from './EditorObject';
+export * from './FabricContext';
